New-style fabric tasks are defined with decorator
A recent version of Fabric introduced a new way to define tasks.
Now, when it loads fabfiles, Fabric mutually excludes either all
tasks that were defined the classic way or all tasks defined the
new-style way.

A recent version of fusionbox-fabric-helpers introduced a new
manner of deploying and staging using the new-style tasks.
Projects that attempt to use the old way of deploying and the new
way of staging (or vice versa) will not able to do both.